benchmarks/eval_benchmarks/safeagentbench/docker/requirements.txt
ai2thor==5.0.0
flask>=3.0
numpy
Pillow
